# docs/04-api-and-integrations/OPENAPI-SPEC.yaml
openapi: 3.0.3
info:
  title: Plataforma de Eventos e Ingressos API
  description: API REST para comercialização de ingressos, controle de assentos sem double-booking, importação de catálogos TMDb/Ticketmaster e validação criptográfica na portaria.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Servidor de Desenvolvimento Local
paths:
...
  /checkout/process:
    post:
      summary: Processa checkout e pagamento simulado (Aprovação / Recusa)
      tags: [Checkout & Pagamento]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reservationId]
              properties:
                reservationId:
                  type: string
                paymentMethod:
                  type: string
                  enum: [SIMULATED_CREDIT_CARD, SIMULATED_PIX]
                  default: SIMULATED_CREDIT_CARD
                action:
                  type: string
                  enum: [APPROVE, REJECT]
                  default: APPROVE
                reason:
                  type: string
                  description: Motivo opcional da recusa simulada
      responses:
        '200':
          description: Pedido processado (retorna Order, Tickets e QR Codes se APPROVED, ou erro e liberação se REJECT)
        '410':
          description: Reserva expirada por tempo limite
...
  /gate/validate:
    post:
      summary: Motor de validação de ingressos na portaria com 4 retornos estritos (QR Code ou Código)
      tags: [Portaria]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId]
              properties:
                eventId:
                  type: string
                  example: "evt_rock2026"
                payload:
                  type: string
                  description: Payload do QR Code com HMAC ou código curto alfanumérico
                  example: "v1:ELT-4819:evt_rock2026:1786579200:e9f1a238b76c8d4e9901ac88f4e2b10a"
                qrPayload:
                  type: string
                  description: Payload completo do QR Code
                ticketCode:
                  type: string
                  description: Código curto alfanumérico do ingresso
      responses:
        '200':
          description: Resultado da validação em um dos 4 estados (VALID, ALREADY_USED, WRONG_EVENT, INVALID_CODE)
          content:
            application/json:
              schema:
                type: object
                required: [result, message]
                properties:
                  result:
                    type: string
                    enum: [VALID, ALREADY_USED, WRONG_EVENT, INVALID_CODE]
                  message:
                    type: string
                  ticket:
                    type: object
...